Restructuring of imports for project models and serializers to allow single model + serializer files to be symlinked into the relevant project folders, and everything (should) then work
