+++ onyx/data/models/__init__.py
@@ -5,9 +5,3 @@ from .models import (
     BaseRecord,
     ProjectRecord,
 )
-from .testmodels import TestModel
-
-try:
-    from .projects import *
-except (ImportError, ModuleNotFoundError):
-    pass
